video_publisher/requirements.txt
Pillow>=5.1.0
confluent-kafka>=0.11.5
keras-applications>=1.0.5
keras-preprocessing>=1.0.3
